PROJECT-89 Move map logic to shared state
