+++ lib/screens/app/profile.dart
@@ -1,8 +1,4 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_sound/flutter_sound.dart';
-import 'package:permission_handler/permission_handler.dart';
 import 'package:project_mobile_app/components/app_bar.dart';
 import 'package:project_mobile_app/state.dart';
 import 'package:provider/provider.dart';
@@ -15,16 +11,6 @@ class ProfileScreen extends StatefulWidget {
 }
 
 class _ProfileScreenState extends State<ProfileScreen> {
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-  }
-
   @override
   Widget build(BuildContext context) {
     var sharedState = Provider.of<SharedState>(context);
