+++ lib/state.dart
@@ -2,33 +2,57 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_sound/flutter_sound.dart';
+import 'package:geolocator/geolocator.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 class SharedState extends ChangeNotifier {
-  String? _username;
   String? _email;
+  String? _username;
+
+  LatLng? _currentLocation;
+
   Duration? _duration;
   double? _decibels;
   double? _maxDecibels;
   LatLng? _maxDecibelsLocation;
-  LatLng? _currentLocation;
 
   final _recorder = FlutterSoundRecorder();
   final _recordingController = StreamController<Food>();
   bool get isRecording => _recorder.isRecording;
 
+  StreamSubscription<Position>? _positionStream;
+  StreamSubscription<RecordingDisposition>? _recordingStream;
+
   SharedState() {
     Permission.microphone.request();
+    _startPositionStream();
   }
 
-  Future startRecording() async {
+  @override
+  void dispose() {
+    stopRecording();
+    _positionStream?.cancel();
+    super.dispose();
+  }
+
+  void _startPositionStream() {
+    _positionStream = Geolocator.getPositionStream().listen((Position pos) {
+      _currentLocation = LatLng(
+        pos.latitude,
+        pos.longitude,
+      );
+      notifyListeners();
+    });
+  }
+
+  void startRecording() async {
     await _recorder.openRecorder();
     await _recorder.setSubscriptionDuration(
       const Duration(milliseconds: 100),
     );
 
-    _recorder.onProgress?.listen((snapshot) {
+    _recordingStream = _recorder.onProgress?.listen((snapshot) {
       _duration = snapshot.duration;
       _decibels = snapshot.decibels!;
 
@@ -46,7 +70,7 @@ class SharedState extends ChangeNotifier {
     notifyListeners();
   }
 
-  Future stopRecording() async {
+  void stopRecording() async {
     await _recorder.stopRecorder();
     await _recorder.closeRecorder();
     notifyListeners();
